Component factory refactor to fix linter issues
